update: extracted types to separate files
