+++ src/Components/QuizCard/QuizCard.types.ts
@@ -1,0 +1,11 @@
+export type CATEGORY = {
+  category: {
+    quizName: string;
+    description: string;
+    questions: {
+      question: string;
+      options: string[];
+      correctOption: string;
+    }[];
+  };
+};
