+++ src/pages/Leaderboard/Leaderboard.types.ts
@@ -1,0 +1,8 @@
+export type SingleLeaderboardEntry = {
+  _id: string;
+  name: string;
+  quizName: string;
+  score: number;
+  __v: number;
+};
+export type LeaderboardData = SingleLeaderboardEntry[];
